BloodBank-RubyonRails
a web application developed using Ruby on Rails which can be used for
the online blood bank service. Interested Donors can register with their
details such as blood group, address and contact. Users seeking blood
can use this application to view the list of donors in their region and
have a contact with them.